requirements.txt
praw~=7.8.1
python-dotenv~=1.1.1
requests~=2.32.4
beautifulsoup4~=4.13.5
scikit-learn~=1.7.1
transformers~=4.56.0
datasets~=4.0.0
accelerate>=1.10.1
fastapi==0.115.7
fastapi-cli==0.0.7
pip~=24.3.1
attrs~=25.3.0
PySocks~=1.7.1
filelock~=3.19.1
dill~=0.3.8
numpy~=2.3.2
pytz~=2025.2
PyYAML~=6.0.2
idna~=3.10
multidict~=6.6.4
propcache~=0.3.2
charset-normalizer~=3.4.3
setuptools~=80.9.0
regex~=2025.7.34
scipy~=1.16.1
threadpoolctl~=3.6.0
typing_extensions~=4.14.1
torch~=2.8.0
mpmath~=1.3.0
sympy~=1.14.0
packaging~=25.0
fsspec~=2025.3.0
aiohttp~=3.12.15
yarl~=1.20.1
MarkupSafe~=3.0.2
joblib~=1.5.2
psutil~=7.0.0
xxhash~=3.5.0
certifi~=2025.8.3
pyarrow~=21.0.0
pandas~=2.3.2
tzdata~=2025.2
urllib3~=2.5.0
six~=1.17.0
python-dateutil~=2.9.0.post0
networkx~=3.5
websocket-client~=1.8.0
safetensors~=0.6.2
multiprocess~=0.70.16